+++ dinky-executor/src/main/java/org/dinky/parser/CreateCDCSourceSqlParser.java
@@ -17,25 +17,25 @@
  *
  */
 
-package org.dinky.parse;
+package org.dinky.parser;
 
 import org.dinky.parser.BaseSingleSqlParser;
 import org.dinky.parser.SqlSegment;
 
 /**
- * ShowFragmentsParser
+ * CreateCDCSourceSqlParser
  *
- * @since 2022/2/17 16:19
+ * @since 2022/1/29 23:39
  */
-public class ShowFragmentParser extends BaseSingleSqlParser {
+public class CreateCDCSourceSqlParser extends BaseSingleSqlParser {
 
-    public ShowFragmentParser(String originalSql) {
+    public CreateCDCSourceSqlParser(String originalSql) {
         super(originalSql);
     }
 
     @Override
     protected void initializeSegments() {
-        // SHOW FRAGMENT (.+)
-        segments.add(new SqlSegment("FRAGMENT", "(show\\s+fragment)\\s+(.*)( ENDOFSQL)", ","));
+        segments.add(new SqlSegment("CDCSOURCE", "(execute\\s+cdcsource\\s+)(.+)(\\s+with\\s+\\()", "[,]"));
+        segments.add(new SqlSegment("WITH", "(with\\s+\\()(.+)(\\))", "',"));
     }
 }
