+++ dinky-executor/src/main/java/org/dinky/trans/ddl/CDCSource.java
@@ -20,7 +20,7 @@
 package org.dinky.trans.ddl;
 
 import org.dinky.assertion.Asserts;
-import org.dinky.parse.SingleSqlParserFactory;
+import org.dinky.parser.SingleSqlParserFactory;
 
 import java.util.ArrayList;
 import java.util.HashMap;
